+++ cx-portal/src/features/admin/idpApiSlice.ts
@@ -26,6 +26,11 @@ export enum IDPCategory {
   OWN = 'OWN',
 }
 
+export enum IDPAuthType {
+  OIDC = 'OIDC',
+  SAML = 'SAML',
+}
+
 export enum IDPMapperType {
   OIDC_USERNAME = 'OIDC_USERNAME',
   HARDCODED_ATTRIBUTE = 'HARDCODED_ATTRIBUTE',
@@ -39,8 +44,17 @@ export enum OIDCAuthMethod {
   SECRET_BASIC = 'SECRET_BASIC',
 }
 
+export enum SAMLAuthMethod {
+  SECRET = 'SECRET',
+}
+
 export interface IDPMapperConfig {
   syncMode: IDPSyncModeType
+}
+
+export interface IDPStatus {
+  id: string
+  enabled: boolean
 }
 
 export interface IDPMapper {
@@ -50,10 +64,16 @@ export interface IDPMapper {
   config: IDPMapperConfig
 }
 
-export interface OIDCType {
+interface BaseAuthType {
   authorizationUrl: string
   clientId: string
+}
+export interface OIDCType extends BaseAuthType {
   clientAuthMethod: OIDCAuthMethod
+}
+
+export interface SAMLType extends BaseAuthType {
+  clientAuthMethod: SAMLAuthMethod
 }
 
 export interface IdentityProvider {
@@ -64,22 +84,49 @@ export interface IdentityProvider {
   redirectUrl: string
   enabled: boolean
   mappers: Array<IDPMapper>
-  oidc: OIDCType
+  oidc?: OIDCType
+  saml?: SAMLType
 }
 
 export const apiSlice = createApi({
   reducerPath: 'rtk/admin/idp',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
   endpoints: (builder) => ({
-    fetchIDPList: builder.query<Array<IdentityProvider>, void>({
-      query: () =>
-        `/api/administration/identityprovider/owncompany/identityproviders`,
+    fetchIDPList: builder.query<Array<IdentityProvider>, number | null>({
+      query: (update: number) =>
+        `/api/administration/identityprovider/owncompany/identityproviders${
+          update ? '?' + update : ''
+        }`,
     }),
     fetchIDPDetail: builder.query<IdentityProvider, string>({
       query: (id: string) =>
         `/api/administration/identityprovider/owncompany/identityproviders/${id}`,
     }),
+    addIDP: builder.mutation<IdentityProvider, IDPAuthType>({
+      query: (protocol: IDPAuthType) => ({
+        url: `/api/administration/identityprovider/owncompany/identityproviders?protocol=${protocol}`,
+        method: 'POST',
+      }),
+    }),
+    removeIDP: builder.mutation<IdentityProvider, string>({
+      query: (id: string) => ({
+        url: `/api/administration/identityprovider/owncompany/identityproviders/${id}`,
+        method: 'DELETE',
+      }),
+    }),
+    enableIDP: builder.mutation<IdentityProvider, IDPStatus>({
+      query: (status: IDPStatus) => ({
+        url: `/api/administration/identityprovider/owncompany/identityproviders/${status.id}/status?enabled=${status.enabled}`,
+        method: 'POST',
+      }),
+    }),
   }),
 })
 
-export const { useFetchIDPListQuery, useFetchIDPDetailQuery } = apiSlice
+export const {
+  useFetchIDPListQuery,
+  useFetchIDPDetailQuery,
+  useAddIDPMutation,
+  useRemoveIDPMutation,
+  useEnableIDPMutation,
+} = apiSlice